+++ client/App.scss
@@ -1,29 +1,5 @@
 .App {
-  text-align: center;
-  height: 100vh;
-  background-color: #555
-}
-
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 80px;
-}
-
-.App-header {
-  height: 150px;
-  padding: 20px;
-  color: white;
-}
-
-.App-title {
-  font-size: 1.5em;
-}
-
-.App-intro {
-  font-size: large;
-}
-
-@keyframes App-logo-spin {
-  from { transform: rotate(0deg); }
-  to { transform: rotate(360deg); }
+	text-align: center;
+	height: 100vh;
+	background-color: #555;
 }
