+++ client/index.tsx
@@ -1,0 +1,20 @@
+import React from 'react'
+import ReactDOM from 'react-dom'
+import App from './App'
+
+import './index.scss'
+
+ReactDOM.render(<App />, document.getElementById('root'))
+
+if ('serviceWorker' in navigator) {
+	window.addEventListener('load', () => {
+		navigator.serviceWorker
+			.register('/service-worker.js')
+			.then(registration => {
+				console.log('SW registered: ', registration)
+			})
+			.catch(registrationError => {
+				console.log('SW registration failed: ', registrationError)
+			})
+	})
+}
