+++ client/pages/page2.tsx
@@ -2,13 +2,13 @@ import React from 'react'
 import { Link } from 'react-router-dom'
 
 const page2 = () => (
-  <div>
-    page 2
-    <br />
-    <button>
-      <Link to="/">Go back</Link>
-    </button>
-  </div>
+	<div>
+		page 2
+		<br />
+		<button>
+			<Link to='/'>Go back</Link>
+		</button>
+	</div>
 )
 
 export default page2
