+++ package.json
@@ -6,6 +6,7 @@
 	"author": "mdfaizan7 <faizaanazim@gmail.com>",
 	"license": "MIT",
 	"dependencies": {
+		"@types/react-router-dom": "^5.1.5",
 		"axios": "^0.19.2",
 		"eslint-config-standard": "^14.1.1",
 		"esm": "^3.2.25",
@@ -32,6 +33,8 @@
 		"@babel/preset-react": "^7.10.4",
 		"@types/express": "^4.17.8",
 		"@types/node": "^14.10.1",
+		"@types/react": "^16.9.49",
+		"@types/react-dom": "^16.9.8",
 		"babel-loader": "^8.1.0",
 		"compression-webpack-plugin": "^5.0.2",
 		"concurrently": "^5.3.0",
